+++ src/components/borderBox1/index.js
@@ -1,4 +1,4 @@
-import React, { useMemo } from 'react'
+import React, { useMemo, forwardRef } from 'react'
 
 import PropTypes from 'prop-types'
 
@@ -7,12 +7,16 @@ import classnames from 'classnames'
 import { deepMerge } from '@jiaminghi/charts/lib/util/index'
 import { deepClone } from '@jiaminghi/c-render/lib/plugin/util'
 
+import useAutoResize from '../../use/autoResize'
+
 import './style.less'
 
 const border = ['left-top', 'right-top', 'left-bottom', 'right-bottom']
 const defaultColor = ['#4fd2dd', '#235fa7']
 
-const BorderBox = ({ children, className, style, color = [] }) => {
+const BorderBox = forwardRef(({ children, className, style, color = [], backgroundColor = 'transparent' }, ref) => {
+  const { width, height, domRef } = useAutoResize(ref)
+
   const mergedColor = useMemo(() => deepMerge(deepClone(defaultColor, true), color || []), [color])
 
   const classNames = useMemo(() => classnames('dv-border-box-1', className), [
@@ -20,7 +24,18 @@ const BorderBox = ({ children, className, style, color = [] }) => {
   ])
 
   return (
-    <div className={classNames} style={style}>
+    <div className={classNames} style={style} ref={domRef}>
+      <svg className='border' width={width} height={height}>
+        <polygon fill={backgroundColor} points={`10, 27 10, ${height - 27} 13, ${height - 24} 13, ${height - 21} 24, ${height - 11}
+        38, ${height - 11} 41, ${height - 8} 73, ${height - 8} 75, ${height - 10} 81, ${height - 10}
+        85, ${height - 6} ${width - 85}, ${height - 6} ${width - 81}, ${height - 10} ${width - 75}, ${height - 10}
+        ${width - 73}, ${height - 8} ${width - 41}, ${height - 8} ${width - 38}, ${height - 11}
+        ${width - 24}, ${height - 11} ${width - 13}, ${height - 21} ${width - 13}, ${height - 24}
+        ${width - 10}, ${height - 27} ${width - 10}, 27 ${width - 13}, 25 ${width - 13}, 21
+        ${width - 24}, 11 ${width - 38}, 11 ${width - 41}, 8 ${width - 73}, 8 ${width - 75}, 10
+        ${width - 81}, 10 ${width - 85}, 6 85, 6 81, 10 75, 10 73, 8 41, 8 38, 11 24, 11 13, 21 13, 24`} />
+      </svg>
+
       {border.map(borderName => (
         <svg
           width='150px'
@@ -66,16 +81,18 @@ const BorderBox = ({ children, className, style, color = [] }) => {
           </polygon>
         </svg>
       ))}
+
       <div className='border-box-content'>{children}</div>
     </div>
   )
-}
+})
 
 BorderBox.propTypes = {
   children: PropTypes.node,
   className: PropTypes.string,
   style: PropTypes.object,
-  color: PropTypes.array
+  color: PropTypes.array,
+  backgroundColor: PropTypes.string
 }
 
 export default BorderBox