+++ src/components/borderBox5/index.js
@@ -13,7 +13,7 @@ import './style.less'
 
 const defaultColor = ['rgba(255, 255, 255, 0.35)', 'rgba(255, 255, 255, 0.20)']
 
-const BorderBox = forwardRef(({ children, reverse = false, className, style, color = [] }, ref) => {
+const BorderBox = forwardRef(({ children, reverse = false, className, style, color = [], backgroundColor = 'transparent' }, ref) => {
   const { width, height, domRef } = useAutoResize(ref)
 
   const mergedColor = useMemo(() => deepMerge(deepClone(defaultColor, true), color || []), [color])
@@ -29,6 +29,9 @@ const BorderBox = forwardRef(({ children, reverse = false, className, style, col
         width={width}
         height={height}
       >
+        <polygon fill={backgroundColor} points={`
+          10, 22 ${width - 22}, 22 ${width - 22}, ${height - 86} ${width - 84}, ${height - 24} 10, ${height - 24}
+        `} />
         <polyline
           className='dv-bb5-line-1'
           stroke={mergedColor[0]}
@@ -73,12 +76,14 @@ BorderBox.propTypes = {
   reverse: PropTypes.bool,
   className: PropTypes.string,
   style: PropTypes.object,
-  color: PropTypes.array
+  color: PropTypes.array,
+  backgroundColor: PropTypes.string
 }
 
 // 指定 props 的默认值：
 BorderBox.defaultProps = {
-  reverse: false
+  reverse: false,
+  backgroundColor: 'transparent'
 }
 
 export default BorderBox